FIX : Renamed Simple Request/Response to Common
