+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/ResultExt.kt
@@ -1,8 +1,6 @@
 package slatekit.common
 
 import slatekit.common.requests.Response
-import slatekit.common.requests.SimpleResponse
-import slatekit.results.*
 
 
 /**
@@ -10,13 +8,13 @@ import slatekit.results.*
  */
 fun <T, E> slatekit.results.Result<T, E>.toResponse(): Response<T> {
     return when (this) {
-        is slatekit.results.Success -> SimpleResponse(this.success, this.code, null, this.value, this.msg, null)
+        is slatekit.results.Success -> CommonResponse(this.success, this.code, null, this.value, this.msg, null)
         is slatekit.results.Failure -> {
             val ex:Exception = when (this.error) {
                 is Exception -> this.error as Exception
                 else -> Exception(this.error.toString())
             }
-            SimpleResponse(this.success, this.code, null, null, this.msg, ex)
+            CommonResponse(this.success, this.code, null, null, this.msg, ex)
         }
     }
 }
