+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/requests/Response.kt
@@ -18,13 +18,6 @@ import slatekit.results.StatusCodes
 /**
  * General purpose class to model a Response at an application boundary ( such as http response )
  * NOTE: This is used for the APIs in Slate Kit
- * @param success : Whether or not the response is successful
- * @param code : A status code ( can be the http status code )
- * @param meta : Meta data for the response ( can be used for headers for http )
- * @param value : The actual value returned by the response
- * @param msg : Message in the case of an failure
- * @param err : Exception in event of failure
- * @param tag : Tag used as a correlation field
  */
 interface Response<out T>  {
     val success: Boolean
