+++ src/lib/kotlin/slatekit-integration/src/main/kotlin/slatekit/integration/apis/SmsApi.kt
@@ -21,7 +21,6 @@ import slatekit.apis.security.Verbs
 import slatekit.apis.support.ApiWithSupport
 import slatekit.common.Context
 import slatekit.common.Vars
-import slatekit.core.common.AppContext
 import slatekit.core.sms.SmsService
 import slatekit.results.Try
