+++ src/lib/kotlin/slatekit-server/src/main/kotlin/slatekit/server/ktor/KtorRequest.kt
@@ -23,13 +23,13 @@ import slatekit.common.*
 import slatekit.server.ServerConfig
 import io.ktor.request.*
 //import kotlinx.coroutines.experimental.async
-import kotlinx.coroutines.async
 import slatekit.common.content.ContentTypeHtml
 import slatekit.common.content.ContentTypeText
 import slatekit.common.content.Doc
 import slatekit.common.requests.Request
 import slatekit.common.requests.RequestSupport
-import slatekit.common.requests.SimpleRequest
+import slatekit.common.CommonRequest
+import slatekit.common.requests.Source
 import java.io.*
 
 class KtorRequest(val call: ApplicationCall, val req: ApplicationRequest) : RequestSupport {
@@ -120,10 +120,10 @@ class KtorRequest(val call: ApplicationCall, val req: ApplicationRequest) : Requ
             // tag    : guid
 
             // Reverting change to args.
-            return SimpleRequest(
+            return CommonRequest(
                     path = uri,
                     parts = parts,
-                    source = slatekit.common.requests.Source.Web,
+                    source = Source.Web,
                     verb = method,
                     meta = KtorHeaders(req, ctx.enc),
                     data = KtorParams(body, req, ctx.enc),
