+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/apis/Api_Core_Tests.kt
@@ -17,9 +17,8 @@ import slatekit.apis.core.Annotated
 import slatekit.apis.core.Api
 import slatekit.apis.security.CliProtocol
 import slatekit.common.auth.Roles
-import slatekit.common.requests.Request
 import slatekit.common.info.Credentials
-import slatekit.common.requests.SimpleRequest
+import slatekit.common.CommonRequest
 import slatekit.common.toResponse
 import slatekit.results.Failure
 import slatekit.results.Success
@@ -38,7 +37,7 @@ class Api_Core_Tests : ApiTestsBase() {
             protocol = CliProtocol,
             apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
             user     = null,
-            request  = SimpleRequest.path("app.users.rolesNone", "get", mapOf(), mapOf(
+            request  = CommonRequest.path("app.users.rolesNone", "get", mapOf(), mapOf(
                     Pair("code", "1"),
                     Pair("tag", "abc")
             )),
@@ -52,7 +51,7 @@ class Api_Core_Tests : ApiTestsBase() {
             protocol = CliProtocol,
             apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
             user     = Credentials(name = "kishore", roles = "dev"),
-            request  = SimpleRequest.path("app.users.argTypeRequest", "get", mapOf(), mapOf(
+            request  = CommonRequest.path("app.users.argTypeRequest", "get", mapOf(), mapOf(
                 Pair("id", "2")
             )),
             response = Success("ok", msg = "raw request id: 2").toResponse()
@@ -65,7 +64,7 @@ class Api_Core_Tests : ApiTestsBase() {
                 protocol = CliProtocol,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "dev"),
-                request  = SimpleRequest.path("app.users.argTypeMeta", "get", mapOf(
+                request  = CommonRequest.path("app.users.argTypeMeta", "get", mapOf(
                         Pair("token", "abc")
                 ), mapOf(
                         Pair("id", "2")
@@ -81,7 +80,7 @@ class Api_Core_Tests : ApiTestsBase() {
                 protocol = CliProtocol,
                 apis     = listOf(Api(SampleErrorsApi(), "app", "sampleErrors", roles = Roles.none, declaredOnly = false)),
                 user     = null,
-                request  = SimpleRequest.path("app.sampleErrors.parseNumberWithResults", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.sampleErrors.parseNumberWithResults", "get", mapOf(), mapOf(
                         "text" to number
                 )),
                 response = Failure( "$number is not a valid number").toResponse()
@@ -94,7 +93,7 @@ class Api_Core_Tests : ApiTestsBase() {
             protocol = CliProtocol,
             apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
             user     = Credentials(name = "kishore", roles = "dev"),
-            request  = SimpleRequest.path("app.users.argTypeListInt", "get", mapOf(), mapOf(
+            request  = CommonRequest.path("app.users.argTypeListInt", "get", mapOf(), mapOf(
                 Pair("items", listOf(1,2,3) )
             )),
             response = Success("ok", msg = ",1,2,3").toResponse()
@@ -107,7 +106,7 @@ class Api_Core_Tests : ApiTestsBase() {
             protocol = CliProtocol,
             apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
             user     = Credentials(name = "kishore", roles = "dev"),
-            request  = SimpleRequest.path("app.users.argTypeListInt", "get", mapOf(), mapOf(
+            request  = CommonRequest.path("app.users.argTypeListInt", "get", mapOf(), mapOf(
                 Pair("items", "1,2,3")
             )),
             response = Success("ok", msg = ",1,2,3").toResponse()
@@ -120,7 +119,7 @@ class Api_Core_Tests : ApiTestsBase() {
             protocol = CliProtocol,
             apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
             user     = Credentials(name = "kishore", roles = "dev"),
-            request  = SimpleRequest.path("app.users.argTypeMapInt", "get", mapOf(), mapOf(
+            request  = CommonRequest.path("app.users.argTypeMapInt", "get", mapOf(), mapOf(
                 Pair("items", mapOf("a" to 1, "b" to 2))
             )),
             response = Success("ok", msg = ",a=1,b=2").toResponse()
@@ -133,7 +132,7 @@ class Api_Core_Tests : ApiTestsBase() {
             protocol = CliProtocol,
             apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
             user     = Credentials(name = "kishore", roles = "dev"),
-            request  = SimpleRequest.path("app.users.argTypeMapInt", "get", mapOf(), mapOf(
+            request  = CommonRequest.path("app.users.argTypeMapInt", "get", mapOf(), mapOf(
                 Pair("items", "a=1,b=2")
             )),
             response = Success("ok", msg = ",a=1,b=2").toResponse()
