+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/apis/Api_Security_Tests.kt
@@ -17,8 +17,7 @@ import slatekit.apis.core.Annotated
 import slatekit.apis.core.Api
 import slatekit.apis.security.AllProtocols
 import slatekit.common.info.Credentials
-import slatekit.common.requests.Request
-import slatekit.common.requests.SimpleRequest
+import slatekit.common.CommonRequest
 import slatekit.common.toResponse
 import slatekit.results.Success
 import slatekit.results.builders.Notices
@@ -38,7 +37,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "dev"),
-                request  = SimpleRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -49,7 +48,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "qa"),
-                request  = SimpleRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -60,7 +59,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = ""),
-                request  = SimpleRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -74,7 +73,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = null,
-                request  = SimpleRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -88,7 +87,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "dev"),
-                request  = SimpleRequest.path("app.users.rolesSpecific", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesSpecific", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -102,7 +101,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "ops"),
-                request  = SimpleRequest.path("app.users.rolesSpecific", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesSpecific", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -116,7 +115,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "admin"),
-                request  = SimpleRequest.path("app.users.rolesParent", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesParent", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -130,7 +129,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "dev"),
-                request  = SimpleRequest.path("app.users.rolesParent", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesParent", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -145,7 +144,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "dev"),
-                request  = SimpleRequest.path("app.users.rolesAny", "get", mapOf(
+                request  = CommonRequest.path("app.users.rolesAny", "get", mapOf(
                         Pair("api-key", "3E35584A8DE0460BB28D6E0D32FB4CFD")
                 ), mapOf(
                         Pair("code", "1"),
@@ -161,7 +160,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = null,
-                request  = SimpleRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
+                request  = CommonRequest.path("app.users.rolesAny", "get", mapOf(), mapOf(
                         Pair("code", "1"),
                         Pair("tag", "abc")
                 )),
@@ -175,7 +174,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "dev"),
-                request  = SimpleRequest.path("app.users.rolesSpecific", "get", mapOf(
+                request  = CommonRequest.path("app.users.rolesSpecific", "get", mapOf(
                         Pair("api-key", "3E35584A8DE0460BB28D6E0D32FB4CFD")
                 ), mapOf(
                         Pair("code", "1"),
@@ -191,7 +190,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "qa"),
-                request  = SimpleRequest.path("app.users.rolesSpecific", "get", mapOf(
+                request  = CommonRequest.path("app.users.rolesSpecific", "get", mapOf(
                         Pair("api-key", "EB7EB37764AD4411A1763E6A593992BD")
                 ), mapOf(
                         Pair("code", "1"),
@@ -207,7 +206,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "admin"),
-                request  = SimpleRequest.path("app.users.rolesParent", "get", mapOf(
+                request  = CommonRequest.path("app.users.rolesParent", "get", mapOf(
                         Pair("api-key", "54B1817194C1450B886404C6BEA81673")
                 ), mapOf(
                         Pair("code", "1"),
@@ -223,7 +222,7 @@ class Api_Security_TestsTests : ApiTestsBase() {
                 protocol = AllProtocols,
                 apis     = listOf(Api(UserApi(ctx), setup = Annotated)),
                 user     = Credentials(name = "kishore", roles = "dev"),
-                request  = SimpleRequest.path("app.users.rolesParent", "get", mapOf(
+                request  = CommonRequest.path("app.users.rolesParent", "get", mapOf(
                         Pair("api-key", "3E35584A8DE0460BB28D6E0D32FB4CFD")
                 ), mapOf(
                         Pair("code", "1"),
