+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/common/DeserializerTests.kt
@@ -1,11 +1,9 @@
 package test.common
 
 import org.junit.Test
-import slatekit.apis.ApiConstants
 import slatekit.common.DateTime
 import slatekit.common.requests.InputArgs
 import slatekit.common.Random
-import slatekit.common.requests.Request
 import slatekit.common.encrypt.EncDouble
 import slatekit.common.encrypt.EncInt
 import slatekit.common.encrypt.EncLong
@@ -15,7 +13,7 @@ import test.setup.Movie
 //import java.time.*
 import org.threeten.bp.*
 import slatekit.common.DateTimes
-import slatekit.common.requests.SimpleRequest
+import slatekit.common.CommonRequest
 import slatekit.common.requests.Source
 import test.setup.MyEncryptor
 import test.setup.StatusEnum
@@ -41,7 +39,7 @@ class ConvertTests {
     fun test_basic_types(tstr:String, tbool:Boolean, tshort:Short, tint:Int, tlong:Long, tdoub:Double):Unit {}
     @Test fun can_parse_basictypes(){
         val test = """{ "tstr": "abc", "tbool": false, "tshort": 1, "tint": 12, "tlong": 123, "tdoub": 123.45 }"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
         val results = deserializer.convert(this::test_basic_types.parameters, test)
         assert(results[0] == "abc")
         assert(results[1] == false)
@@ -55,7 +53,7 @@ class ConvertTests {
     fun test_dates(tdate: LocalDate, ttime: LocalTime, tlocaldatetime: LocalDateTime, tdatetime: DateTime):Unit{}
     @Test fun can_parse_dates(){
         val test = """{ "tdate": "2017-07-06", "ttime": "10:30:45", "tlocaldatetime": "2017-07-06T10:30:45", "tdatetime": "201707061030" }"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
         val results = deserializer.convert(this::test_dates.parameters, test)
         assert(results[0] == LocalDate.of(2017, 7, 6))
         assert(results[1] == LocalTime.of(10,30,45))
@@ -68,7 +66,7 @@ class ConvertTests {
     @Test fun can_parse_guids(){
         val guid = Random.guid()
         val test = """{ "uid": "$guid" }"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
         val results = deserializer.convert(this::test_uuid.parameters, test)
         assert(results[0] == UUID.fromString(guid))
     }
@@ -78,7 +76,7 @@ class ConvertTests {
     @Test fun can_parse_enum(){
         val enumVal = StatusEnum.Active
         val test = """{ "status": ${enumVal.value} }"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
         val results = deserializer.convert(this::test_enum.parameters, test)
         assert(results[0] == enumVal)
     }
@@ -93,7 +91,7 @@ class ConvertTests {
         val decDoub = MyEncryptor.encrypt("12345.67")
 
         val test = """{ "decString": "$decStr", "decInt": "$decInt", "decLong": "$decLong", "decDouble": "$decDoub" }"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()), MyEncryptor)
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()), MyEncryptor)
         val results = deserializer.convert(this::test_decrypted.parameters, test)
         assert((results[0] as EncString).value == "abc123")
         assert((results[1] as EncInt).value == 123)
@@ -105,7 +103,7 @@ class ConvertTests {
     fun test_arrays(strings: List<String>, bools:List<Boolean>, ints:List<Int>, longs:List<Long>, doubles:List<Double>):Unit {}
     @Test fun can_parse_arrays(){
         val test = """{ "strings": ["a", "b", "c"], "bools": [true, false, true], "ints": [1,2,3], "longs": [100,200,300], "doubles": [1.2,3.4,5.6] }"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()), MyEncryptor)
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()), MyEncryptor)
         val results = deserializer.convert(this::test_arrays.parameters, test)
         assert((results[0] as List<String>)[0] == "a")
         assert((results[0] as List<String>)[1] == "b")
@@ -129,7 +127,7 @@ class ConvertTests {
     fun test_object(sample1: SampleObject1):Unit{}
     @Test fun can_parse_object(){
         val test = """{ "sample1": { "tstr": "abc", "tbool": false, "tshort": 1, "tint": 12, "tlong": 123, "tdoub": 123.45 } }"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
         val results = deserializer.convert(this::test_object.parameters, test)
         assert(results[0] == ConvertTests.SampleObject1("abc", false, 1, 12, 123, 123.45))
     }
@@ -142,7 +140,7 @@ class ConvertTests {
             { "tstr": "abc", "tbool": false, "tshort": 1, "tint": 12, "tlong": 123, "tdoub": 123.45 },
             { "tstr": "def", "tbool": true , "tshort": 2, "tint": 34, "tlong": 456, "tdoub": 678.91 }
         ]}"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
         val inputs = deserializer.convert(this::test_object_list.parameters, test)
         val results = inputs.get(0) as ArrayList<*>
         println(results)
@@ -166,7 +164,7 @@ class ConvertTests {
                 ]
             }
         }"""
-        val deserializer = Deserializer(SimpleRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
+        val deserializer = Deserializer(CommonRequest.cli("a", "b", "c", "post", mapOf(), mapOf()))
         val results = deserializer.convert(this::test_nested_object_list.parameters, test)
         val item = results[1] as NestedObject1
         assert(results[0] == "abc")
@@ -178,7 +176,7 @@ class ConvertTests {
     fun test_custom_converter(tstr:String, tbool:Boolean, movie: Movie):Unit {}
     @Test fun can_parse_custom_types(){
         val test = """{ "tstr": "abc", "tbool": false }"""
-        val req = SimpleRequest("a.b.c", listOf("a", "b", "c"), Source.CLI, "post",
+        val req = CommonRequest("a.b.c", listOf("a", "b", "c"), Source.CLI, "post",
                 InputArgs(mapOf()), InputArgs(mapOf(Pair("movie", "batman"))))
         val deserializer = Deserializer( req,null, mapOf(Pair(Movie::class.qualifiedName!!, { request, json, tpe ->
             Movie(0L, request.meta.getString("movie"), cost = 0, rating = 4.0, released = DateTime.now() )
