+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/common/HealthTests.kt
@@ -4,7 +4,7 @@ import org.junit.Assert
 import org.junit.Test
 import slatekit.integration.common.Health
 import slatekit.common.info.Build
-import slatekit.core.common.AppContext
+import slatekit.common.CommonContext
 
 class HealthTests {
 
@@ -12,7 +12,7 @@ class HealthTests {
     @Test
     fun can_get_version() {
         val build = Build("1.2.3.4", "abc123", "master", "2018-11-20")
-        val health = Health(AppContext.simple("test").copy(build = build))
+        val health = Health(CommonContext.simple("test").copy(build = build))
         val info = health.version()
         Assert.assertEquals("version", info.name)
         Assert.assertEquals("health", info.source)
@@ -26,7 +26,7 @@ class HealthTests {
     @Test
     fun can_get_info() {
         val build = Build("1.2.3.4", "abc123", "master", "2018-11-20")
-        val health = Health(AppContext.simple("test").copy(build = build))
+        val health = Health(CommonContext.simple("test").copy(build = build))
         val info = health.info()
         println(info)
     }
