+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/workers/Worker_System_Tests.kt
@@ -3,7 +3,7 @@ package test.workers
 import org.junit.Assert
 import slatekit.common.TODO
 import slatekit.common.metrics.MetricsLite
-import slatekit.core.common.AppContext
+import slatekit.common.CommonContext
 import slatekit.workers.*
 import slatekit.common.Status
 
@@ -19,7 +19,7 @@ class Worker_System_Tests {
 
     fun build():System {
         TODO.IMPLEMENT("tests", "Workers")
-        val sys = System(AppContext.simple("test"), listOf(), metrics = MetricsLite.build())
+        val sys = System(CommonContext.simple("test"), listOf(), metrics = MetricsLite.build())
         TODO.IMPLEMENT("tests", "Workers")
 //        sys.register("group 1", Worker("group1", "worker 1",  callback = { success("worktype 1:done") } ))
 //        sys.register("group 1", Worker("group1", "worker 2",  callback = { success("worktype 2:done") } ))
